requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.1
pillow==11.3.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-learn==1.7.1
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.168
ultralytics-thop==2.0.14
urllib3==2.5.0
